+++ src/Plugin.php
@@ -2,6 +2,9 @@
 
 namespace craft\awss3;
 
+use craft\base\Element;
+use craft\elements\Asset;
+use craft\events\ModelEvent;
 use craft\events\RegisterComponentTypesEvent;
 use craft\services\Volumes;
 use yii\base\Event;
@@ -36,6 +39,33 @@ class Plugin extends \craft\base\Plugin
 
         Event::on(Volumes::class, Volumes::EVENT_REGISTER_VOLUME_TYPES, function(RegisterComponentTypesEvent $event) {
             $event->types[] = Volume::class;
+        });
+
+        Event::on(Asset::class, Element::EVENT_AFTER_SAVE, function (ModelEvent $event) {
+            /** @var Asset $asset */
+            $asset = $event->sender;
+
+            /** @var Volume $volume */
+            $volume = $asset->getVolume();
+
+            if (!$volume instanceof Volume) {
+                return;
+            }
+
+            if (!$volume->autoFocalPoint) {
+                return;
+            }
+
+            $fullPath = (!empty($volume->subfolder) ? rtrim($volume->subfolder, '/') . '/'  : '') . $asset->getPath();
+
+            $focalPoint = $volume->detectFocalPoint($fullPath);
+
+            if (!empty($focalPoint)) {
+                $assetRecord = \craft\records\Asset::findOne($asset->id);
+                $assetRecord->focalPoint = implode(';', $focalPoint);
+                $assetRecord->save();
+            }
+
         });
     }
 }