+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Fixed
 - Fixed an error where it was not possible to use this plugin on installs that were updated from Craft 2 to Craft 3.1 ([#38](https://github.com/craftcms/aws-s3/issues/38))
+- Fixed a migration error. ([#39](https://github.com/craftcms/aws-s3/issues/39))
 
 ## 1.0.8 - 2018-01-02